# elderly-care-backend/docs/swagger.yaml
definitions:
...
  models.LocationUpdate:
    properties:
      accuracy:
        type: number
      address:
        type: string
      heading:
        type: number
      latitude:
        type: number
      longitude:
        type: number
      speed:
        type: number
      user_id:
        type: integer
    type: object
  models.UserLocation:
    properties:
      accuracy:
        description: 定位精度
        type: number
      address:
        type: string
      createdAt:
        type: string
      heading:
        description: 移动方向
        type: number
      id:
        description: 自增ID
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      speed:
        description: 移动速度
        type: number
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  vo.ResponseVO:
    properties:
      data: {}
      msg:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 养老服务平台接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: elderly-care-backend
  version: "1.0"
paths:
...
  /location/history:
    get:
      consumes:
      - application/json
      description: 获取用户的最近位置记录
      parameters:
      - default: 10
        description: 返回数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 获取用户历史位置
      tags:
      - 定位模块
  /location/navigation:
    get:
      consumes:
      - application/json
      description: 根据起点终点计算步行导航路线
      parameters:
      - description: 起点坐标 格式:经度,纬度
        in: query
        name: from
        required: true
        type: string
      - description: 终点坐标 格式:经度,纬度
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      summary: 计算导航路线
      tags:
      - 定位模块
  /location/navigation/location:
    get:
      consumes:
      - application/json
      description: 根据位置记录ID计算导航路线
      parameters:
      - description: 位置记录ID
        in: query
        name: locationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 根据位置ID导航
      tags:
      - 定位模块
  /location/navigation/to-target:
    get:
      consumes:
      - application/json
      description: 获取当前用户到指定目标的导航路线
      parameters:
      - description: 目标纬度
        in: query
        name: targetLat
        required: true
        type: number
      - description: 目标经度
        in: query
        name: targetLng
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 获取用户到目标的导航路线
      tags:
      - 定位模块
  /location/navigation/user:
    get:
      consumes:
      - application/json
      description: 计算两个用户之间的导航路线
      parameters:
      - description: 目标用户ID
        in: query
        name: targetUserId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 用户间导航
      tags:
      - 定位模块
  /location/nearby:
    get:
      consumes:
      - application/json
      description: 获取附近的志愿者或求助者
      parameters:
      - description: 纬度
        in: query
        name: lat
        required: true
        type: number
      - description: 经度
        in: query
        name: lng
        required: true
        type: number
      - default: 5000
        description: 搜索半径
        in: query
        name: radius
        type: number
      - description: 用户角色
        enum:
        - volunteer
        - user
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.ResponseVO'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取附近用户
      tags:
      - 实时定位
  /location/reverse-geocode:
    get:
      consumes:
      - application/json
      description: 将经纬度坐标转换为详细地址
      parameters:
      - description: 经度
        in: query
        name: lng
        required: true
        type: string
      - description: 纬度
        in: query
        name: lat
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      summary: 坐标转地址
      tags:
      - 定位模块
  /location/update:
    post:
      consumes:
      - application/json
      description: 实时更新用户当前位置
      parameters:
      - description: 位置更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LocationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 更新用户位置
      tags:
      - 实时定位
  /location/user/{userId}:
    get:
      consumes:
      - application/json
      description: 获取指定用户的当前位置
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.ResponseVO'
            - properties:
                data:
                  $ref: '#/definitions/models.UserLocation'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户当前位置
      tags:
      - 实时定位
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"